history/ClickHouse/requirements.txt
certifi==2018.4.16
chardet==3.0.4
funcsigs==1.0.2
future==0.16.0
idna==2.6
infi.clickhouse-orm==0.9.8
iso8601==0.1.12
jsonrpcclient==2.5.2
jsonrpcserver==3.5.4
jsonschema==2.6.0
pkg-resources==0.0.0
pymongo==3.6.1
pytz==2018.4
requests==2.18.4
six==1.11.0
tornado==5.0.2
urllib3==1.22
clickhouse-driver
